Modularized & tested working
