+++ c_src/_exponential_ld.c
@@ -18,18 +18,7 @@
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <Python.h>
 #include "numpy/arrayobject.h"
-#include <math.h>
-
-#if defined (_OPENMP)
-#  include <omp.h>
-#endif
-
-#ifndef M_PI
-    #define M_PI 3.14159265358979323846
-#endif
-
-#define MIN(x, y) (((x) < (y)) ? (x) : (y))
-#define MAX(x, y) (((x) > (y)) ? (x) : (y))
+#include "common.h"
 
 static PyObject *_exponential_ld(PyObject *self, PyObject *args);
 
@@ -41,26 +30,13 @@ double intensity(double x, double c1, double c2, double norm)
 	return (1. - c1*(1. - mu) - c2/(1. - exp(mu)))/norm; 
 }
 
-double area(double d, double x, double R)
-{
-	/*
-	Returns area of overlapping circles with radii r and R; separated by a distance d
-	*/
-	double arg1 = (d*d + x*x - R*R)/(2.*d*x); 	
-	double arg2 = (d*d + R*R - x*x)/(2.*d*R); 
-	double arg3 = MAX((-d + x + R)*(d + x - R)*(d - x + R)*(d + x + R), 0.);
-
-	if(x <= R - d) return M_PI*x*x;						//planet completely overlaps stellar circle
-	else if(x >= R + d) return M_PI*R*R;					//stellar circle completely overlaps planet
-	else return x*x*acos(arg1) + R*R*acos(arg2) - 0.5*sqrt(arg3);		//partial overlap
-}
 
 static PyObject *_exponential_ld(PyObject *self, PyObject *args)
 {
-	double rprs, d, fac, A_i, x, I; 
+	double rprs, fac;
 	int nthreads;
-	npy_intp i, dims[1];
-	double dx, A_f, x_in, x_out, delta, c1, c2;
+	npy_intp dims[1];
+	double c1, c2;
 	
 	PyArrayObject *ds, *flux;
   	if(!PyArg_ParseTuple(args,"Oddddi", &ds, &rprs, &c1, &c2, &fac, &nthreads)) return NULL;
@@ -84,50 +60,10 @@ static PyObject *_exponential_ld(PyObject *self, PyObject *args)
 		Laura Kreidberg 07/2015
 	*/
 	
-	#if defined (_OPENMP)
-	omp_set_num_threads(nthreads);	//specifies number of threads (if OpenMP is supported)
-	#endif
-
 	double norm = 2.*M_PI*(0.5 - 0.1666666667*c1 + 0.77750463*c2); 	//normalization for intensity profile (faster to calculate it once, rather than every time intensity is called)		
-
-	#if defined (_OPENMP)
-	#pragma omp parallel for private(d, x_in, x_out, delta, x, dx, A_i, A_f, I)
-	#endif
-	for(i = 0; i < dims[0]; i++)
-	{
-		d = d_array[i];
-		x_in = MAX(d - rprs, 0.);					//lower bound for integration
-		x_out = MIN(d + rprs, 1.0);					//upper bound for integration
-
-		if(x_in >= 1.) f_array[i] = 1.0;				//flux = 1. if the planet is not transiting
-		else
-		{
-			delta = 0.;						//variable to store the integrated intensity, \int I dA
-			x = x_in;						//starting radius for integration
-			dx = fac*acos(x); 					//initial step size 
-
-			x += dx;						//first step
-
-			A_i = 0.;						//initial area
+	double intensity_args[] = {c1, c2, norm};
+	calc_limb_darkening(f_array, d_array, dims[0], rprs, fac, nthreads, intensity, intensity_args);
 	
-			while(x < x_out)
-			{
-				A_f = area(d, x, rprs);				//calculates area of overlapping circles
-				I = intensity(x - dx/2., c1, c2, norm); 	//intensity at the midpoint
-				delta += (A_f - A_i)*I;				//increase in transit depth for this integration step
-				dx = fac*acos(x);  				//updating step size
-				x = x + dx;					//stepping to next element
-				A_i = A_f;					//storing area
-			}
-			dx = x_out - x + dx;  					//calculating change in radius for last step  
-			x = x_out;						//final radius for integration
-			A_f = area(d, x, rprs);					//area for last integration step
-			I = intensity(x - dx/2., c1, c2, norm); 		//intensity at the midpoint 
-			delta += (A_f - A_i)*I;					//increase in transit depth for this integration step
-
-			f_array[i] = 1.0 - delta;	//flux equals 1 - \int I dA 
-		}
-	}
 	return PyArray_Return((PyArrayObject *)flux);
 
 }
