+++ c_src/_nonlinear_ld.c
@@ -18,46 +18,22 @@
 #define NPY_NO_DEPRECATED_API NPY_1_7_API_VERSION
 #include <Python.h>
 #include "numpy/arrayobject.h"
+#include "common.h"
 
-#if defined (_OPENACC) && defined(__PGI)
-#  include <accelmath.h>
-#else
-#  include <math.h>
-#endif
-
-#if defined (_OPENMP)
-#  include <omp.h>
-#endif
-
-#ifndef M_PI
-    #define M_PI 3.14159265358979323846
-#endif
 
 #define MIN(x, y) (((x) < (y)) ? (x) : (y))
 #define MAX(x, y) (((x) > (y)) ? (x) : (y))
 
 static PyObject *_nonlinear_ld(PyObject *self, PyObject *args);
 
-inline double intensity(double x, double c1, double c2, double c3, double c4, double norm)
+double intensity(double x, double* args)
 {
+	//args should contain c1, c2, c3, c4, norm, in that order
 	if(x > 0.99995) x = 0.99995;
 	double sqrtmu = pow(1. - x*x,0.25);
-	return (1. - c1*(1. - sqrtmu) - c2*(1. - pow(sqrtmu,2.)) - c3*(1. - pow(sqrtmu, 3.)) - c4*(1. - pow(sqrtmu,4.)))/norm; 	
+	return (1. - args[0]*(1. - sqrtmu) - args[1]*(1. - pow(sqrtmu,2.)) - args[2]*(1. - pow(sqrtmu, 3.)) - args[3]*(1. - pow(sqrtmu,4.)))/args[4];
 }
 
-inline double area(double d, double x, double R)
-{
-	/*
-	Returns area of overlapping circles with radii r and R; separated by a distance d
-	*/
-	double arg1 = (d*d + x*x - R*R)/(2.*d*x); 	
-	double arg2 = (d*d + R*R - x*x)/(2.*d*R); 
-	double arg3 = MAX((-d + x + R)*(d + x - R)*(d - x + R)*(d + x + R), 0.);
-
-	if(x <= R - d) return M_PI*x*x;						//planet completely overlaps stellar circle
-	else if(x >= R + d) return M_PI*R*R;					//stellar circle completely overlaps planet
-	else return x*x*acos(arg1) + R*R*acos(arg2) - 0.5*sqrt(arg3);		//partial overlap
-}
 
 static PyObject *_nonlinear_ld(PyObject *self, PyObject *args)
 {
@@ -74,6 +50,7 @@ static PyObject *_nonlinear_ld(PyObject *self, PyObject *args)
 
 	double *f_array = PyArray_DATA(flux);
 	double *d_array = PyArray_DATA(ds);
+	double norm = (-c1/10. - c2/6. - 3.*c3/14. - c4/4. + 0.5)*2.*M_PI;
 
 	/*
 		NOTE:  the safest way to access numpy arrays is to use the PyArray_GETITEM and PyArray_SETITEM functions.
@@ -87,53 +64,8 @@ static PyObject *_nonlinear_ld(PyObject *self, PyObject *args)
 
 		Laura Kreidberg 07/2015
 	*/
-	
-	#if defined (_OPENMP) && !defined(_OPENACC)
-	omp_set_num_threads(nthreads);	//specifies number of threads (if OpenMP is supported)
-	#endif
-
-	double norm = (-c1/10. - c2/6. - 3.*c3/14. - c4/4. + 0.5)*2.*M_PI; 	//normalization for intensity profile (faster to calculate it once, rather than every time intensity is called)		
-
-	#if defined (_OPENACC)
-	#pragma acc parallel loop copyout(f_array[:dims[0]])
-	#elif defined (_OPENMP)
-	#pragma omp parallel for
-	#endif
-	for(int i = 0; i < dims[0]; i++)
-	{
-		double d = d_array[i];
-		double x_in = MAX(d - rprs, 0.);					//lower bound for integration
-		double x_out = MIN(d + rprs, 1.0);					//upper bound for integration
-
-		if(x_in >= 1.) f_array[i] = 1.0;				//flux = 1. if the planet is not transiting
-		else
-		{
-			double delta = 0.;						//variable to store the integrated intensity, \int I dA
-			double x = x_in;						//starting radius for integration
-			double dx = fac*acos(x); 					//initial step size
-
-			x += dx;						//first step
-
-			double A_i = 0.;						//initial area
-	
-			while(x < x_out)
-			{
-				double A_f = area(d, x, rprs);				//calculates area of overlapping circles
-				double I = intensity(x - dx/2.,c1,c2, c3, c4, norm); 	//intensity at the midpoint
-				delta += (A_f - A_i)*I;				//increase in transit depth for this integration step
-				dx = fac*acos(x);  				//updating step size
-				x = x + dx;					//stepping to next element
-				A_i = A_f;					//storing area
-			}
-			dx = x_out - x + dx;  					//calculating change in radius for last step  FIXME
-			x = x_out;						//final radius for integration
-			double A_f = area(d, x, rprs);					//area for last integration step
-			double I = intensity(x - dx/2.,c1,c2, c3, c4, norm); 		//intensity at the midpoint
-			delta += (A_f - A_i)*I;					//increase in transit depth for this integration step
-
-			f_array[i] = 1.0 - delta;	//flux equals 1 - \int I dA 
-		}
-	}
+	double intensity_args[] = {c1, c2, c3, c4, norm};
+	calc_limb_darkening(f_array, d_array, dims[0], rprs, fac, nthreads, intensity, intensity_args);
 	return PyArray_Return((PyArrayObject *)flux);
 
 }
