+++ setup.py
@@ -64,12 +64,12 @@ has_openmp, needs_gomp = detect_openmp()
 parallel_args = ['-fopenmp'] if has_openmp else []
 parallel_libraries = ['gomp'] if needs_gomp else []
 
-_nonlinear_ld = Extension('batman._nonlinear_ld', ['c_src/_nonlinear_ld.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
+_nonlinear_ld = Extension('batman._nonlinear_ld', ['c_src/_nonlinear_ld.c', 'c_src/common.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
 _quadratic_ld = Extension('batman._quadratic_ld', ['c_src/_quadratic_ld.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
-_uniform_ld   = Extension('batman._uniform_ld', ['c_src/_uniform_ld.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
-_logarithmic_ld   = Extension('batman._logarithmic_ld', ['c_src/_logarithmic_ld.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
-_exponential_ld   = Extension('batman._exponential_ld', ['c_src/_exponential_ld.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
-_custom_ld   = Extension('batman._custom_ld', ['c_src/_custom_ld.c', 'c_src/_custom_intensity.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
+_uniform_ld   = Extension('batman._uniform_ld', ['c_src/_uniform_ld.c', 'c_src/common.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
+_logarithmic_ld   = Extension('batman._logarithmic_ld', ['c_src/_logarithmic_ld.c', 'c_src/common.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
+_exponential_ld   = Extension('batman._exponential_ld', ['c_src/_exponential_ld.c', 'c_src/common.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
+_custom_ld   = Extension('batman._custom_ld', ['c_src/_custom_ld.c', 'c_src/common.c'], extra_compile_args = parallel_args, libraries = parallel_libraries) 
 _rsky = Extension('batman._rsky', ['c_src/_rsky.c'])
 _eclipse = Extension('batman._eclipse', ['c_src/_eclipse.c'])
